--- apps/backend/src/app/analytics/repository/sales-analytics-repository.interface.ts
@@ -1,11 +1,0 @@
-export const SALES_ANALYTICS_REPOSITORY = Symbol('SALES_ANALYTICS_REPOSITORY');
-
-export interface ISalesAnalyticsRepository {
-    fetchMonthlyProductQuantity(params: {
-        fromYYYYMM: string;
-        toYYYYMM: string;
-        timezone: string;
-    }): Promise<any[]>;
-
-    
-}
