+++ apps/backend/src/app/analytics/sales-analytics/custom/DTOs/base/index.ts
