+++ apps/backend/src/app/analytics/sales-analytics/custom/DTOs/base/monthly-profit.dto.ts
