+++ apps/backend/src/app/analytics/sales-analytics/custom/DTOs/base/monthly-quantity.dto.ts
