+++ apps/backend/src/app/analytics/sales-analytics/custom/DTOs/request/index.ts
@@ -1,0 +1,1 @@
+export * from './top-products-request.dto';
