+++ apps/backend/src/app/analytics/sales-analytics/custom/DTOs/request/top-products-request.dto.ts
