+++ apps/backend/src/app/analytics/sales-analytics/custom/DTOs/response/index.ts
@@ -1,0 +1,2 @@
+export * from './top-products-profit-response.dto';
+export * from './top-products-quantity-response.dto';
