+++ apps/backend/src/app/analytics/sales-analytics/custom/DTOs/response/top-products-profit-response.dto.ts
