+++ apps/backend/src/app/analytics/sales-analytics/custom/DTOs/response/top-products-quantity-response.dto.ts
