+++ apps/backend/src/app/analytics/sales-analytics/custom/mappers/monthly-profit.mapper.ts
