+++ apps/backend/src/app/analytics/sales-analytics/custom/mappers/monthly-quantity.mapper.ts
@@ -1,0 +1,6 @@
+import { MonthlyQuantity } from '@common/Interfaces';
+import { MonthlyQuantityDto } from '../DTOs/base/monthly-quantity.dto';
+
+export function mapMonthlyQuantity(rows: MonthlyQuantity[]): MonthlyQuantityDto[] {
+    return rows.map((row) => new MonthlyQuantityDto(row.month, row.productId, row.quantity));
+}
