+++ apps/backend/src/app/analytics/sales-analytics/custom/repository/sales-custom-repository.interface.ts
@@ -1,0 +1,5 @@
+export const SALES_ANALYTICS_REPOSITORY = Symbol('SALES_ANALYTICS_REPOSITORY');
+
+export interface ISalesAnalyticsRepository {
+    fetchMonthlyProductQuantity(params: { fromYYYYMM: string; toYYYYMM: string; timezone: string }): Promise<any[]>;
+}
