+++ apps/backend/src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { MongooseModule } from '@nestjs/mongoose';
-import { SalesAnalyticsModule } from './analytics/sales-analytics.module';
+import { SalesAnalyticsModule } from './analytics/sales-analytics/sales-analytics.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { AuthModule } from './auth/auth.module';
